remove one more warning
git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.2/program@598 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_graph_helper.c
@@ -119,7 +119,7 @@ rrd_parse_color(char *string, graph_desc_t *gdp) {
     */
 
     i=0;
-    while (string[i] && isxdigit(string[i])) i++;
+    while (string[i] && isxdigit((unsigned int)string[i])) i++;
     if (string[i] != '\0') return 1; /* garbage follows hexdigits */
     switch (i) {
 	case 3:
